# backend/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ficha de Treino API",
    "version": "0.1.0",
    "description": "API REST do sistema Ficha de Treino. Endpoints de autenticação, treinos, exercícios e gerenciamento de conta."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor da API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/workouts/{weekDay}/exercises": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Adiciona um exercício ao treino diário do usuário autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weekDay",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DOMINGO",
                "SEGUNDA",
                "TERCA",
                "QUARTA",
                "QUINTA",
                "SEXTA",
                "SABADO"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "exerciseId"
                ],
                "properties": {
                  "exerciseId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Associação criada com done inicialmente falso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "exerciseId",
                    "done"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "exerciseId": {
                      "type": "string"
                    },
                    "done": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisição inválido"
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado"
          },
          "404": {
            "description": "Dia, treino ou exercício não encontrado"
          },
          "409": {
            "description": "Exercício já existe no treino"
          }
        }
      }
    },
    "/api/workouts/{weekDay}/exercises/{exerciseId}": {
      "delete": {
        "tags": [
          "Workouts"
        ],
        "summary": "Remove um exercício do treino diário do usuário autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weekDay",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DOMINGO",
                "SEGUNDA",
                "TERCA",
                "QUARTA",
                "QUINTA",
                "SEXTA",
                "SABADO"
              ]
            }
          },
          {
            "in": "path",
            "name": "exerciseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identificador do exercício no catálogo"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercício removido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "deleted"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado"
          },
          "404": {
            "description": "Dia inválido, associação inexistente ou pertencente a outro usuário"
          }
        }
      }
    },
...
    "/api/exercises": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Busca exercícios por nome e metadados",
        "description": "A busca ignora diferenças entre maiúsculas, minúsculas e acentos; filtros repetidos usam OR e filtros diferentes usam AND.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Termo de busca; vazio retorna o catálogo a partir do primeiro exercício"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "alongamento",
                  "cardio",
                  "forca",
                  "levantamento-olimpico",
                  "pliometria",
                  "powerlifting",
                  "strongman"
                ]
              }
            },
            "style": "form",
            "explode": true,
            "description": "Uma ou mais categorias"
          },
          {
            "in": "query",
            "name": "equipment",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "barra",
                  "barra-w",
                  "bola-de-exercicio",
                  "bola-medicinal",
                  "cabo",
                  "faixas",
                  "halteres",
                  "kettlebell",
                  "maquina",
                  "outros",
                  "peso-do-corpo",
                  "rolo-de-espuma"
                ]
              }
            },
            "style": "form",
            "explode": true,
            "description": "Um ou mais equipamentos"
          },
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "avancado",
                  "iniciante",
                  "intermediario"
                ]
              }
            },
            "style": "form",
            "explode": true,
            "description": "Um ou mais níveis"
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "pull",
                  "push",
                  "static"
                ]
              }
            },
            "style": "form",
            "explode": true,
            "description": "Um ou mais tipos de força"
          },
          {
            "in": "query",
            "name": "mechanic",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "composto",
                  "isolado"
                ]
              }
            },
            "style": "form",
            "explode": true,
            "description": "Um ou mais tipos de mecânica"
          },
          {
            "in": "query",
            "name": "primaryMuscle",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "abdominais",
                  "abdutores",
                  "adutores",
                  "antebracos",
                  "biceps",
                  "dorsais",
                  "gluteos",
                  "inferior-das-costas",
                  "isquiotibiais",
                  "meio-das-costas",
                  "ombros",
                  "panturrilhas",
                  "peito",
                  "pescoco",
                  "quadriceps",
                  "trapezio",
                  "triceps"
                ]
              }
            },
            "style": "form",
            "explode": true,
            "description": "Um ou mais músculos principais"
          },
          {
            "in": "query",
            "name": "secondaryMuscle",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "abdominais",
                  "abdutores",
                  "adutores",
                  "antebracos",
                  "biceps",
                  "dorsais",
                  "gluteos",
                  "inferior-das-costas",
                  "isquiotibiais",
                  "meio-das-costas",
                  "ombros",
                  "panturrilhas",
                  "peito",
                  "pescoco",
                  "quadriceps",
                  "trapezio",
                  "triceps"
                ]
              }
            },
            "style": "form",
            "explode": true,
            "description": "Um ou mais músculos secundários"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercícios correspondentes à busca e total antes da paginação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "isCustom",
                          "primaryMuscles",
                          "secondaryMuscles",
                          "instructions",
                          "images"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "force": {
                            "type": "string",
                            "nullable": true
                          },
                          "isCustom": {
                            "type": "boolean"
                          },
                          "level": {
                            "type": "string",
                            "nullable": true
                          },
                          "mechanic": {
                            "type": "string",
                            "nullable": true
                          },
                          "equipment": {
                            "type": "string",
                            "nullable": true
                          },
                          "primaryMuscles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "secondaryMuscles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "instructions": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "category": {
                            "type": "string",
                            "nullable": true
                          },
                          "images": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros de busca ou paginação inválidos"
          },
          "401": {
            "description": "Token JWT ausente, inválido ou expirado"
          }
        }
      }
    },
...
  },
...
}